requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
requests>=2.26.0
pyinstaller>=5.0.0
tkcalendar>=1.6.1 